Display of gameboard extracted to method
note: As a teacher sometiems it's best to let your students
make the mistakes.  That way they learn from it and we
all have a good laugh.

note: and sometimes the teacher likes to giggle as a student
does not spot the obvious bug for ten minutes because she
'knew' where to look for it.